git-svn-id: http://citadel.physiology.northwestern.edu/svn/s1_analysis@838 984b1e3c-1c57-0410-ae62-c18f90d3b645

+++ proc/robert/predictions_mwstikpolyMOD.m
@@ -119,7 +119,7 @@ end
 
 if ~strcmpi(signal,'emg')
 q = find_active_regions_words(bdf.words,bdf.vel(:,1));
-q = interp1(bdf.vel(:,1), q, t);
+q = interp1(bdf.vel(:,1), double(q), t);
 else
     q=ones(size(t))';
 end
